+++ python/cugraph/structure/graph_primtypes.pxd
@@ -214,7 +214,7 @@ cdef extern from "utilities/cython.hpp" namespace "cugraph::cython":
         numberTypeEnum vertexType,
         numberTypeEnum edgeType,
         numberTypeEnum weightType,
-        int num_partition_edges,
+        size_t num_partition_edges,
         size_t num_global_vertices,
         size_t num_global_edges,
         size_t row_comm_size,
